requirements.txt
selenium>=4.0.0
bibtexparser>=1.4.0
requests>=2.28.0
python-dotenv>=1.0.0